requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
arrow==1.2.1
    # via django-q
asgiref==3.4.1
    # via django
authlib==0.15.5
    # via -r requirements.in
backports.zoneinfo==0.2.1
    # via django
beautifulsoup4==4.10.0
    # via django-bootstrap4
blessed==1.19.0
    # via django-q
brotli==1.0.9
    # via whitenoise
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.10
    # via requests
cryptography==36.0.1
    # via
    #   authlib
    #   msal
    #   pyjwt
cycler==0.11.0
    # via matplotlib
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
django==4.0.1
    # via
    #   -r requirements.in
    #   django-bootstrap4
    #   django-health-check
    #   django-import-export
    #   django-localflavor
    #   django-multiselectfield
    #   django-phonenumber-field
    #   django-picklefield
    #   django-q
django-bootstrap4==21.2
    # via -r requirements.in
django-countries==7.2.1
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-health-check==3.16.5
    # via -r requirements.in
django-import-export==2.7.1
    # via -r requirements.in
django-localflavor==3.1
    # via -r requirements.in
django-multiselectfield==0.1.12
    # via -r requirements.in
django-phonenumber-field==6.0.0
    # via -r requirements.in
django-picklefield==3.0.1
    # via django-q
django-q==1.3.9
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
fonttools==4.28.5
    # via matplotlib
idna==3.3
    # via requests
kiwisolver==1.3.2
    # via matplotlib
ldap3==2.9.1
    # via -r requirements.in
markuppy==1.14
    # via tablib
matplotlib==3.5.1
    # via -r requirements.in
msal==1.16.0
    # via -r requirements.in
numpy==1.22.1
    # via matplotlib
oauthlib==3.1.1
    # via requests-oauthlib
odfpy==1.4.1
    # via tablib
openpyxl==3.0.9
    # via tablib
packaging==21.3
    # via matplotlib
phonenumbers==8.12.41
    # via -r requirements.in
pillow==9.0.0
    # via matplotlib
pyasn1==0.4.8
    # via ldap3
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via msal
pyparsing==3.0.6
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   arrow
    #   matplotlib
python-stdnum==1.17
    # via django-localflavor
pyyaml==6.0
    # via tablib
redis==3.5.3
    # via django-q
requests==2.27.1
    # via
    #   -r requirements.in
    #   msal
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via -r requirements.in
six==1.16.0
    # via
    #   blessed
    #   python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.1.0
    # via django-import-export
urllib3==1.26.8
    # via requests
wcwidth==0.2.5
    # via blessed
whitenoise[brotli]==5.3.0
    # via -r requirements.in
xlrd==2.0.1
    # via
    #   -r requirements.in
    #   tablib
xlsxwriter==3.0.2
    # via -r requirements.in
xlwt==1.3.0
    # via tablib
